funcionando eliminar

+++ src/main/java/com/dh/odontologica/service/OdontologoService.java
@@ -1,5 +1,6 @@
 package com.dh.odontologica.service;
 
+import com.dh.odontologica.exceptions.ResourceNotFoundException;
 import com.dh.odontologica.model.OdontologoDTO;
 import com.dh.odontologica.persistence.entity.Odontologo;
 import com.dh.odontologica.persistence.repository.OdontologoRepository;
@@ -30,8 +31,14 @@ public class OdontologoService {
         return respuesta;
     }
 
-    public void eliminar(Long id){
-        repository.deleteById(id);
+    public void eliminar(Long id)throws ResourceNotFoundException {
+
+        if(buscarPorId(id).isEmpty()) {
+            throw new ResourceNotFoundException("No existe este registro con el id" + " " + id);
+        }else {
+            repository.deleteById(id);
+        }
+
     }
 
     public Optional<Odontologo> buscarPorId(Long id){
